feat(es6) : switch a new batch of modules to ES6
Refs: https://github.com/yargs/yargs/issues/2112
