+++ package.json
@@ -49,15 +49,17 @@
   "author": "Ben Coe <ben@npmjs.com>",
   "license": "ISC",
   "dependencies": {
-    "string-width": "^4.2.0",
-    "strip-ansi": "^6.0.1",
-    "wrap-ansi": "^7.0.0"
+    "@rollup/plugin-commonjs": "^22.0.0",
+    "@types/wrap-ansi": "^8.0.1",
+    "string-width": "^5.1.2",
+    "strip-ansi": "^7.0.1",
+    "wrap-ansi": "^8.0.1"
   },
   "devDependencies": {
+    "@rollup/plugin-node-resolve": "^13.3.0",
     "@types/node": "^14.0.27",
     "@typescript-eslint/eslint-plugin": "^4.0.0",
     "@typescript-eslint/parser": "^4.0.0",
-    "@wessberg/rollup-plugin-ts": "^1.3.2",
     "c8": "^7.3.0",
     "chai": "^4.2.0",
     "chalk": "^4.1.0",
@@ -69,6 +71,7 @@
     "mocha": "^9.1.3",
     "rimraf": "^3.0.2",
     "rollup": "^2.23.1",
+    "rollup-plugin-ts": "^2.0.7",
     "standardx": "^7.0.0",
     "typescript": "^4.0.0"
   },
